version 0.1-16

+++ src/SYMPHONY/CoinUtils/src/CoinSearchTree.hpp
@@ -1,4 +1,8 @@
-/* $Id: CoinSearchTree.hpp 1191 2009-07-25 08:38:12Z forrest $ */
+/* $Id: CoinSearchTree.hpp 1372 2011-01-03 23:31:00Z lou $ */
+// Copyright (C) 2006, International Business Machines
+// Corporation and others.  All Rights Reserved.
+// This code is licensed under the terms of the Eclipse Public License (EPL).
+
 #ifndef CoinSearchTree_H
 #define CoinSearchTree_H
 
@@ -335,13 +339,13 @@ protected:
     }
     /** After changing data in the top node, fix the heap */
     virtual void fixTop() {
-	const int size = candidateList_.size();
+	const size_t size = candidateList_.size();
 	if (size > 1) {
 	    CoinTreeSiblings** candidates = &candidateList_[0];
 	    CoinTreeSiblings* s = candidates[0];
 	    --candidates;
-	    int pos = 1;
-	    int ch;
+	    size_t pos = 1;
+	    size_t ch;
 	    for (ch = 2; ch < size; pos = ch, ch *= 2) {
 		if (comp_(candidates[ch+1], candidates[ch]))
 		    ++ch;
@@ -362,8 +366,8 @@ protected:
 	candidateList_.push_back(s);
 	CoinTreeSiblings** candidates = &candidateList_[0];
 	--candidates;
-	int pos = candidateList_.size();
-	int ch;
+	size_t pos = candidateList_.size();
+	size_t ch;
 	for (ch = pos/2; ch != 0; pos = ch, ch /= 2) {
 	    if (comp_(candidates[ch], s))
 		break;
